requirements.txt
Django==1.5.1
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6
South==0.8.1
Sphinx==1.2b1
bpython==0.12
coverage==3.6
distribute==0.6.49
django-ajax-selects==1.3.1
django-braces==1.2.2
django-crispy-forms==1.4.0
django-crispy-forms-fancy-formsets==0.1
django-debug-toolbar==0.9.4
django-discover-runner==0.4
django-grappelli==2.4.7
django-model-utils==1.4.0
django-selectable==0.7.0
django-suit==0.2.5
django-taggit==0.10
django-tastypie==0.10.0
django-uuidfield==0.4.0
djangorestframework==2.3.9
docutils==0.11
gunicorn==0.17.4
logutils==0.3.3
mimeparse==0.1.3
python-dateutil==2.2
simplejson==3.3.1
six==1.4.1
stevedore==0.10
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
wsgiref==0.1.2
xlrd==0.9.2
